"Manually" load modules as plugins
Do not use the plugin registry to load the modules at runtime.
