+++ plugins/traction_innkeeper/traction_innkeeper/v1_0/schema_storage/routes.py
@@ -53,7 +53,7 @@ class SchemaStorageAddSchema(OpenAPISchema):
     schema_id = fields.Str(description="Schema identifier", required=True)
 
 
-class OperationResponseSchema(OpenAPISchema):
+class SchemaStorageOperationResponseSchema(OpenAPISchema):
     """Response schema for simple operations."""
 
     success = fields.Bool(
@@ -118,7 +118,7 @@ async def schema_storage_get(request: web.BaseRequest):
     tags=[SWAGGER_CATEGORY],
 )
 @match_info_schema(SchemaIdMatchInfoSchema())
-@response_schema(OperationResponseSchema(), 200, description="")
+@response_schema(SchemaStorageOperationResponseSchema(), 200, description="")
 @error_handler
 async def schema_storage_remove(request: web.BaseRequest):
     context: AdminRequestContext = request["context"]
